+++ src/main/java/pg/provingground/service/CarService.java
@@ -58,13 +58,13 @@ public class CarService {
     }
 
     /** 입력된 연료량을 유효한 값으로 변경하여 반환 */
-    public int validFuelAmount(Long carId, int amount) {
+    public Long validFuelAmount(Long carId, Long amount) {
         if (amount <= 0) {
             throw new IllegalArgumentException("유효하지 않은 주유량입니다.");
         }
         Car car = carRepository.findOne(carId);
         // 연료 탱크의 남은 공간
-        int leftCapacity = car.getType().getFuelCapacity() - car.getFuel();
+        Long leftCapacity = car.getType().getFuelCapacity() - car.getFuel();
         return Math.min(amount * 1000, leftCapacity) ;
     }
